+++ LMS/Models/LMSModels/Enrolled.cs
@@ -6,7 +6,7 @@ namespace LMS.Models.LMSModels
     public partial class Enrolled
     {
         public string UId { get; set; }
-        public short ClassId { get; set; }
+        public int ClassId { get; set; }
         public string Grade { get; set; }
 
         public virtual Classes Class { get; set; }
