+++ LMS/Models/LMSModels/Submissions.cs
@@ -6,7 +6,7 @@ namespace LMS.Models.LMSModels
     public partial class Submissions
     {
         public string UId { get; set; }
-        public short AssId { get; set; }
+        public int AssId { get; set; }
         public DateTime? Time { get; set; }
         public string Contents { get; set; }
         public uint? Score { get; set; }
